+++ src/env_forked.json
@@ -1,0 +1,6 @@
+{
+  "rpcHost" : "https://test.icomonitor.io/nodes/mainnet-fork",
+  "GaTrackingId" : "",
+  "apiUrl": "https://platform.neufund.net",
+  "fbPixelId": "000000000"
+}
